+++ conf/topsites_global.csv
@@ -1,3 +1,4 @@
+rank,site
 1,google.com
 2,facebook.com
 3,youtube.com
