+++ lib/__about__.py
